Move class Net::LDAP::Connection to a separate file.
